// web/console/lib/ff-redux/src/base/Model/ActionTypesEnum.ts
export enum ActionTypesEnum {
  ChangeKeyword = 'ChangeKeyword',
  PerformSearch = 'PerformSearch',

  ChangeContinueToken = 'ChangeContinueToken',

  ChangeFilter = 'ChangeFilter',
  ApplyFilter = 'ApplyFilter',
  ApplyPolling = 'ApplyPolling',

  ChangeSearchFilter = 'ChangeSearchFilter',
  ApplySearchFilter = 'ApplySearchFilter',

  ChangePaging = 'ChangePaging',
  ChangePagingIndex = 'ChangePagingIndex',
  RestPaging = 'RestPaging',
  Next = 'Next',

  SortBy = 'SortBy',
  Reset = 'Reset'
}